src/templates/grappelli/requirements.txt
Django==1.11.21
django-grappelli==2.12.3
django-filebrowser==3.11.1
django-tinymce==2.8.0
Pillow==6.0.0
